agent/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
apscheduler
beautifulsoup4
feedparser
fyers-apiv3
joblib
numpy
openai
pandas
pandas_market_calendars
psycopg2-binary
pyotp
python-dotenv
requests
scikit-learn
supabase
torch
transformers
xgboost
yfinance
SQLAlchemy
lxml
